Start moving dart:ui into //flutter/lib/ui
This patch moves the geometry classes.
