+++ flutter/lib/ui/dart_ui.gni
@@ -1,0 +1,10 @@
+# Copyright 2016 The Chromium Authors. All rights reserved.
+# Use of this source code is governed by a BSD-style license that can be
+# found in the LICENSE file.
+
+dart_ui_files = [
+  "//flutter/lib/ui/geometry.dart",
+  "//flutter/lib/ui/ui.dart",
+]
+
+dart_ui_path = "//flutter/lib/ui/ui.dart"
