+++ sky/engine/core/core.gni
@@ -224,15 +224,8 @@ core_dart_files = get_path_info([
                                   "dart/window.dart",
                                   "painting/Color.dart",
                                   "painting/FilterQuality.dart",
-                                  "painting/Offset.dart",
-                                  "painting/OffsetBase.dart",
                                   "painting/Paint.dart",
                                   "painting/PaintingStyle.dart",
-                                  "painting/Point.dart",
-                                  "painting/RRect.dart",
-                                  "painting/RSTransform.dart",
-                                  "painting/Rect.dart",
-                                  "painting/Size.dart",
                                   "painting/TransferMode.dart",
                                 ],
                                 "abspath")
