+++ travis/analyze.sh
@@ -10,7 +10,7 @@ RESULTS=`dartanalyzer                                                          \
   --strong-hints                                                               \
   --fatal-hints                                                                \
   --lints                                                                      \
-  out/host_debug_unopt/gen/sky/bindings/dart_ui.dart                         \
+  out/host_debug_unopt/gen/sky/bindings/dart_ui/ui.dart                        \
   2>&1                                                                         \
   | grep -v "\[error\] Target of URI does not exist: 'dart:mojo.internal'"     \
   | grep -v "\[error\] Native functions can only be declared in the SDK and code that is loaded through native extensions" \
@@ -20,10 +20,10 @@ RESULTS=`dartanalyzer                                                          \
   | grep -v "\[warning\] Undefined name 'MojoHandleWatcher'"                   \
   | grep -v "\[warning\] Undefined name 'MojoCoreNatives'"                     \
   | grep -v "\[error\] The library ''dart:_internal'' is internal"             \
-  | grep -Ev "Unused import .+dart_ui\.dart"                                   \
+  | grep -Ev "Unused import .+ui\.dart"                                        \
   | grep -v "\[info\] TODO"                                                    \
   | grep -Ev "[0-9]+ errors.*found."                                           \
-  | grep -v "Analyzing \[out/host_debug_unopt/gen/sky/bindings/dart_ui.dart\]\.\.\."`
+  | grep -v "Analyzing \[out/host_debug_unopt/gen/sky/bindings/dart_ui/ui.dart\]\.\.\."`
 
 echo "$RESULTS"
 if [ -n "$RESULTS" ];
